optimize

+++ internal/model/system/response/sys_user.go
@@ -11,5 +11,5 @@ type SysUserResponse struct {
 type LoginResponse struct {
 	User      system.SysUser `json:"user"`
 	Token     string         `json:"token"`
-	ExpiresAt int64          `json:"expiresAt"`
+	ExpiresAt float64        `json:"expiresAt"`
 }
